+++ CHANGELOG.md
@@ -7,23 +7,40 @@
 - Using `rescale` as augmentation is not allowed anymore as the rescaling of input
   images will now be done via the keras.application `preprocess_input' function of the
   backbone model used (#314).
+- Support for all optimizers offered by keras was added, but the setting has now become 
+  case sensitive. For Keras >= 3, AdamW is now the default optimizer
+  (#334, #335, #338).
 
 ### Improvements
 
 - Add support to reproject a local image layer (#290)
+- Add support to configure an image layer based on a simple dir with images (#337)
 - Improve handling of `switch_axes` and `ssl_verify` (#295)
 - Add proper support for keras 3:
     - Use `categorical_focal_crossentropy` by default on keras 3 (#312)
     - Use [segmodels_keras](https://github.com/orthoseg/segmodels_keras) (#311)
     - Avoid unneeded direct use of tensorflow (#304, #313)
+    - Use ``segmodels_keras.utils.load_weights`` for forward and backward compatibility
+      when loading cross-keras-version weights (#339)
 - Use keras.applications `preprocess_input` functions to prepare inputs (#314)
+- Significant improvements to the documentation (#316, #330,...)
+
+## 0.7.1 (2026-04-13)
+
+### Improvements
+
 - Add retry for ARCGIS WMS server errors (#305)
 - Add retry when reading vrt layers (#307)
+- Improve handling of read errors during prediction (#325)
+- Use small mobilenetv2 model instead of the default in footballfield sample project
+  (#130)
+- Don't download sample projects from git repo (#323)
 
 ### Bugs fixed
 
 - Fix use of username/password for WMS image layer (#294)
 - Fix models not being saved if architecture_id != 1 (#310)
+- Fix some issues with handing ~ in input paths (#321)
 
 ## 0.7.0 (2025-08-01)